Text.RE2.Types

Description

Types used in compiling and matching regexes.

Regular expressions

data RE2 enc

Abstract type representing a compiled regex.

Parametrized by a character encoding, either UTF8 or Latin1.

data UTF8

Tag type denoting a regex which operates on UTF8-encoded text.

data Latin1

Tag type denoting a regex which operates on Latin1-encoded text.

Compile options

data CompileOption

Options available when compiling a regex.

Defaults (in parentheses) are according to http://code.google.com/p/re2/source/browse/re2/re2.h. The Haskell binding does not enforce these.

If multiple options with the same constructor are given in a list, the last one has priority.

Constructors

PosixSyntax Bool
(False) Restrict regexps to POSIX egrep syntax.

LongestMatch Bool
(False) Search for longest match, not first match.

LogErrors Bool
(True) Print errors to stderr.

Literal Bool
(False) Interpret string as literal, not regexp.

NeverNewline Bool
(False) Never match \n, even if it is in regexp.

CaseSensitive Bool
(True) Match is case-sensitive.

PerlClasses Bool
Allow Perl's \d \s \w \D \S \W even in POSIX mode.

WordBoundary Bool
Allow Perl's \b \B even in POSIX mode.

OneLine Bool
Match ^ and $ only at beginning and end of text even in POSIX mode.

MaxMemory Int
Memory limit, in bytes.

Match options

data MatchOptions

Options available when matching with a regex.

For default options use defMatchOptions, or import Data.Default and use def.

Constructors

MatchOptions

Fields

anchor :: Anchor

Anchor match at start or end of input?

captureLimit :: Maybe Int

If Just n, only store the first n capturing groups. This includes the whole-match group, so Just 1 returns only that group, and Just 0 is a Boolean match only.

defMatchOptions :: MatchOptions

Default MatchOptions.

data Anchor

Is a match required to coincide with the start or end of input?

Constructors

Unanchored
No anchor; match may occur anywhere inside input.

AnchorStart
Anchor at start only.

AnchorBoth
Anchor at start and end.

Results

data Error

An error which occurred while compiling a regex.

Constructors

Error

Fields

errCode :: Maybe Int

Errors which originate in RE2 itself have a numeric code. For now, the meaning of these codes is not exposed in Haskell.

errMessage :: String

Human-readable description of the error.

errFragment :: ByteString

A fragment of the regex relating to the error. When the regex was given as String or Text, this fragment will be UTF-8.

newtype Match str

A successful regex match.

Groups which did not participate in this match are represented as Nothing.

Constructors

Match (Seq (Maybe (Group str)))

data Group str

A capturing group.

Parametrized by the string type: ByteString, Text, or String.

Constructors

Group

Fields

byteStart :: Int

The byte index of the start of this group.

byteLength :: Int

The length of this group in bytes.

captured :: str

The string which was captured.

data Stats

Some statistics about a compiled regex.

Constructors

Stats

Fields

numGroups :: Int

The number of capturing groups, not including the whole-match group.

programSize :: Int

The program size, a rough cost measure.